Refacto and new high-level functions
