+++ demos/multigrid/samurai_new/indices.hpp
@@ -1,10 +1,83 @@
 #pragma once
-#include "stencil.hpp"
+#include <samurai/algorithm.hpp>
 
 using namespace samurai;
 
 namespace samurai_new
 {
+    /**
+     * Stores the triplet (level, i, index)
+    */
+    template<class Mesh>
+    struct MeshInterval
+    {
+        using interval_t = typename Mesh::interval_t;
+        using coord_type = typename Mesh::lca_type::iterator::coord_type;
+
+        std::size_t level;
+        interval_t i;
+        coord_type index;
+        double cell_length;
+
+        MeshInterval(std::size_t l) 
+        : level(l) 
+        {
+            cell_length = 1./(1 << level);
+        }
+
+        MeshInterval(std::size_t l, const interval_t& _i, const coord_type& _index) 
+        : level(l) , i(_i), index(_index)
+        {
+            cell_length = 1./(1 << level);
+        }
+    };
+
+    template <class Mesh>
+    inline auto get_index_start(const Mesh& mesh, const MeshInterval<Mesh>& mesh_interval)
+    {
+        static constexpr std::size_t dim = Mesh::dim;
+        using coord_index_t = typename Mesh::coord_index_t;
+
+        std::array<coord_index_t, dim> coord;
+        std::copy(mesh_interval.index.cbegin(), mesh_interval.index.end(), coord.begin()+1);
+        coord[0] = mesh_interval.i.start;
+        return mesh.get_index(mesh_interval.level, coord);
+    }
+
+    template <class Mesh, class Vector>
+    inline auto get_index_start_translated(const Mesh& mesh, const MeshInterval<Mesh>& mesh_interval, const Vector& translation_vect)
+    {
+        static constexpr std::size_t dim = Mesh::dim;
+        using coord_index_t = typename Mesh::coord_index_t;
+
+        std::array<coord_index_t, dim> coord;
+        std::copy(mesh_interval.index.cbegin(), mesh_interval.index.end(), coord.begin()+1);
+        coord[0] = mesh_interval.i.start;
+        for (std::size_t d=0; d<dim; ++d)
+        {
+            coord[d] += translation_vect[d];
+        }
+        return mesh.get_index(mesh_interval.level, coord);
+    }
+
+    template <class Mesh, class Vector>
+    inline auto get_index_start_children(const Mesh& mesh, MeshInterval<Mesh>& mesh_interval, const Vector& translation_vect)
+    {
+        static constexpr std::size_t dim = Mesh::dim;
+        using coord_index_t = typename Mesh::coord_index_t;
+
+        std::array<coord_index_t, dim> coord;
+        std::copy(mesh_interval.index.cbegin(), mesh_interval.index.cend(), coord.begin()+1);
+        coord[0] = mesh_interval.i.start;
+        for (std::size_t d=0; d<dim; ++d)
+        {
+            coord[d] = 2*coord[d] + translation_vect[d];
+        }
+
+        return mesh.get_index(mesh_interval.level+1, coord);
+    }
+
+
     //
     // Functions that return cell indices
     //
@@ -28,6 +101,13 @@ namespace samurai_new
         });
     }
 
+    template <typename DesiredIndexType, class Mesh, class Func>
+    inline void for_each_cell(const Mesh& mesh, Func &&f)
+    {
+        using mesh_id_t = typename Mesh::mesh_id_t;
+        for_each_cell<DesiredIndexType>(mesh, mesh[mesh_id_t::cells], std::forward<Func>(f));
+    }
+
     template <typename DesiredIndexType, class Mesh, class Subset, class Func>
     inline void for_each_cell(const Mesh& mesh, std::size_t level, Subset& subset, Func &&f)
     {
@@ -39,39 +119,6 @@ namespace samurai_new
             for_each_cell<DesiredIndexType>(mesh, mesh_interval, std::forward<Func>(f));
         });
     }
-
-    template <typename DesiredIndexType, class Mesh, std::size_t stencil_size, class Func>
-    inline void for_each_stencil(const Mesh& mesh, const MeshInterval<Mesh>& mesh_interval, StencilIndices<DesiredIndexType, stencil_size, Mesh::dim>& stencil, Func &&f)
-    {
-        stencil.init(mesh, mesh_interval);
-        f(stencil.indices());
-        for(DesiredIndexType ii=1; ii<static_cast<DesiredIndexType>(mesh_interval.i.size()); ++ii)
-        {
-            stencil.move_next();
-            f(stencil.indices());
-        }
-    }
-
-    template <typename DesiredIndexType, class Mesh, class Set, std::size_t stencil_size, class Func>
-    inline void for_each_stencil(const Mesh& mesh, const Set& set, std::size_t level, StencilIndices<DesiredIndexType, stencil_size, Mesh::dim>& stencil, Func &&f)
-    {
-        MeshInterval<Mesh> mesh_interval(level);
-        for_each_interval(set[level], [&](std::size_t /*level*/, const auto& i, const auto& index)
-        {
-            mesh_interval.i = i;
-            mesh_interval.index = index;
-            for_each_stencil<DesiredIndexType>(mesh, mesh_interval, stencil, std::forward<Func>(f));
-        });
-    }
-
-    /*template <typename DesiredIndexType, class Mesh, class Set, class StencilType, class Func>
-    inline void for_each_stencil(const Mesh& mesh, const Set& set, StencilType& stencil, Func &&f)
-    {
-        for_each_level(set, [&](std::size_t level)
-        {
-            for_each_stencil<DesiredIndexType>(mesh, set, level, stencil, std::forward<Func>(f));
-        });
-    }*/
 
 
 
@@ -105,25 +152,6 @@ namespace samurai_new
             cell.indices[0]++;
             cell.index++;
         }
-    }
-
-    template <class Mesh, std::size_t stencil_size, class Func>
-    inline void for_each_stencil(const Mesh& mesh, const MeshInterval<Mesh>& mesh_interval, StencilCells<Mesh, stencil_size>& stencil, Func &&f)
-    {
-        stencil.init(mesh, mesh_interval);
-        f(stencil.cells());
-        for(std::size_t ii=1; ii<mesh_interval.i.size(); ++ii)
-        {
-            stencil.move_next();
-            f(stencil.cells());
-        }
-    }
-
-    template <class Mesh, std::size_t stencil_size, class Func>
-    inline void for_each_stencil(const Mesh& mesh, const MeshInterval<Mesh>& mesh_interval, const Stencil<stencil_size, Mesh::dim>& stencil_shape, Func &&f)
-    {
-        StencilCells<Mesh, stencil_size> stencil(stencil_shape);
-        for_each_stencil(mesh, mesh_interval, stencil, std::forward<Func>(f));
     }
